third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_pointerrawupdate_flush_pointercapture.https-expected.txt

This is a testharness.js-based test.
[FAIL] Pointer capture state should be updated before `pointerrawupdate`
  assert_equals: expected "pointerrawupdate@target, pointermove@target, pointerdown@target, gotpointercapture@target, pointerrawupdate@target, pointermove@target, pointerup@target, lostpointercapture@target, pointerover@dest" but got "pointerrawupdate@target, pointermove@target, pointerdown@target, pointerrawupdate@target, gotpointercapture@target, pointermove@target, pointerup@target, lostpointercapture@target, pointerover@dest"
Harness: the test ran to completion.
